requirements.txt
fastapi==0.95.2
uvicorn==0.30.6
jinja2==3.1.4
pyyaml==6.0.2
pymongo==4.6.2
speedtest-cli==2.1.3
